# api-spec-v3.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/matthewminish/seng302-2021-api-spec/3.0.1
info:
  description: API specification for the SENG302 2021 project
  version: 3.0.1
  title: Wasteless
  contact:
    email: seng302@canterbury.ac.nz
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
security:
  - CookieAuth:
      - user
      - globalApplicationAdmin
      - defaultGlobalApplicationAdmin
paths:
...
  '/listings/{listingId}/notify':
    post:
      tags:
        - 'Notifications'
        - 'U31 : Purchases'
      summary: Notify a user about any changes of items they liked or purchased.
      security:
        - CookieAuth:
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - in: path
          name: listingId
          schema:
            type: integer
          description: Id of the listing that has been updated
          required: true
          example: 54343
      responses:
        '201':
          description: Successful notification
        '401':
          description: Cannot send a notification to a user when the sender is
            not logged in.
        '406':
          description: Listing ID not found
  '/users/{userId}/notifications':
    get:
      tags:
        - 'Notifications'
        - 'U31 : Purchases'
      summary: User can retrieve all notifications related to listings they have received.
      security:
        - CookieAuth:
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          description: Id of the user who is about to receive the notification
          required: true
          example: 111
      responses:
        '200':
          description: Successful retrieval of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListOfNotifications'
        '401':
          description: Cannot send a notification to a user when the receiver is
            not logged in.
        '400':
          description: Bad request
        '406':
          description: User ID not found
  '/businesses/{businessId}/notifications':
    get:
      tags:
        - 'Listings / sales'
        - 'U31 : Purchases'
      summary: Business can retrieve all notifications related to listings they have sold.
      security:
        - CookieAuth:
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - in: path
          name: businessId
          schema:
            type: integer
          description: Id of the business who is about to receive the notification
          required: true
          example: 111
      responses:
        '200':
          description: Successful retrieval of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListOfNotifications'
        '401':
          description: Cannot send a notification to a user when the receiver is
            not logged in.
        '400':
          description: Bad request
        '406':
          description: Business ID not found
  '/notifications/{notificationId}':
    delete:
      tags:
        - 'U32 : Managing my feed'
        - 'Notifications'
      summary: 'Delete a notification for a user'
      parameters:
        - name: notificationId
          in: path
          description: Id of the notification to delete
          required: true
          schema:
            type: integer
            example: 1
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '200':
          description: Request successful, notification is deleted.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: 'Forbidden: Returned when a user tries to delete a notification that does not belong to them AND the user is not a GAA'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
    put:
      tags:
        - 'U32 : Managing my feed'
        - 'Notifications'
      summary: Update the view status of a user's notification
      parameters:
        - name: notificationId
          in: path
          description: Id of the notification to update
          required: true
          schema:
            type: integer
            example: 1
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                viewStatus:
                  type: string
                  description: The new viewing status to change the notification to
                  example: "READ"
      responses:
        '200':
          description: Update successful, notification status was updated
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: 'Forbidden: Returned when a user tries to update a notification that does not belong to them AND the user is not a GAA'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
...
  '/businesses/listings/{id}/like':
    post:
      tags:
        - 'Listings / sales'
        - 'U29 : Browse/search sale listings'
      summary: Add a user like to a sale listing
      security:
        - CookieAuth:
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: Id of the listing
          required: true
          example: 50
      responses:
        '201':
          description: The listing was successfully liked
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
    delete:
      tags:
        - 'Listings / sales'
        - 'U29 : Browse/search sale listings'
      summary: Delete a user like on a listing
      security:
        - CookieAuth:
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: Id of the listing
          required: true
          example: 50
      responses:
        '200':
          description: The listing like was successfully removed
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
  '/users/{id}/likes':
    get:
      tags:
        - 'Users'
        - 'U29 : Browse/search sale listings'
      summary: Retrieved all liked listings for a specific user
      security:
        - CookieAuth:
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: Id of the user to obtain likes from
          required: true
          example: 3
      responses:
        '200':
          description: A (potentially empty) list of liked listings was retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Listing'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    BadRequestError:
      description: There was some error with the data supplied by the user, appropriate error message(s) should be shown to user
    UnauthorizedError:
      description: Access token is missing or invalid
...
    NotAcceptableError:
      description: 'The requested route does exist (so not a 404) but some part of the request is not acceptable, for example trying to access a resource by an ID that does not exist.'
...
  schemas:
...
    Listing:
      type: object
      description: A listing for the sale of some product from a business's inventory
      required:
        - inventoryItemId
        - quantity
        - price
      properties:
        id:
          type: integer
          example: 57
          readOnly: true
          description: A unique (within the business) identifer for this listing - autogenerated by the system
        inventoryItem:
          readOnly: true
          allOf:
            - $ref: '#/components/schemas/InventoryItem'
        inventoryItemId:
          type: integer
          example: 101
          writeOnly: true
        quantity:
          type: integer
          example: 3
          description: Must be greater than zero, and less than or equal to the quantity of the correpsonding inventory entry
        price:
          type: number
          format: double
          description: Price of the listing
          example: 17.99
        moreInfo:
          type: string
          example: 'Seller may be willing to consider near offers'
          description: An optional field for further information about the listing
        created:
          type: string
          format: date-time
          example: '2021-07-14T11:44:00Z'
          readOnly: true
          description: Date and time when the listing was created
        closes:
          type: string
          format: date-time
          example: '2021-07-21T23:59:00Z'
          description: Date and time when the listing closes (defaults to expiry date of related inventory item)
        likes:
          type: number
          format: integer
          example: 10
          description: The number of likes the listing has received (defaults to 0)
    InventoryItem:
      type: object
      description: Entity representing an instance of a stock of some product
      required:
        - productId
        - quantity
        - expires
      properties:
        id:
          type: integer
          example: 101
          readOnly: true
          description: A unique (within the inventory) identifer for this item - autogenerated by the system
        product:
          readOnly: true
          allOf:
            - $ref: '#/components/schemas/Product'
        productId:
          writeOnly: true
          type: string
          example: "WATT-420-BEANS"
          description: Product Code of corresponding product in business's catalogue. Notice that the ID is given when writing, but the full object is returned when reading
        quantity:
          type: integer
          example: 4
          description: How many items of stock are present
        pricePerItem:
          type: number
          format: double
          description: Price to purchase a single item of the product
          example: 6.50
        totalPrice:
          type: number
          format: double
          description: Price for purchasing the entire stock of the product
          example: 21.99
        manufactured:
          type: string
          format: date
          description: The date on which the product was prepared / manufactured
        sellBy:
          type: string
          format: date
          description: The date by which the product ought to be sold
        bestBefore:
          type: string
          format: date
          description: 'The "best before" date of the product'
        expires:
          type: string
          format: date
          description: The expiry date of the product
    Image:
      type: object
      description: Entity representing an image stored on the server
      required:
        - filename
        - thumbnailFilename
      properties:
        id:
          type: integer
          example: 1234
          readOnly: true
          description: A unique identifier (system-wide) to represent this image
        filename:
          type: string
          example: "/media/images/23987192387509-123908794328.png"
          description: Relative URI to full size image. A fully qualified URI (e.g www.wasteless.com/media/images/...) should NOT be given.
        thumbnailFilename:
          type: string
          example: "/media/images/23987192387509-123908794328_thumbnail.png"
          description: Relative URI to thumbnail of image. A fully qualified URI (e.g www.wasteless.com/media/images/...) should NOT be given.
    Product:
      type: object
      description: Entity for catalogue items in a product catalogue
      required:
        - name
      properties:
        id:
          type: string
          example: 'WATT-420-BEANS'
          description: >
            Developers can choose whether or not to make this a required field (i.e user
            specified vs. auto-generated). In the case that the team automatically
            generates the IDs, they should still allow this field to be present when
            creating (POST) or modifying (PUT) a product, and then simply ignore its
            value (i.e do not return a 400 code if it is present but not needed). An
            ID only needs to be unique within its catalogue, not system-wide.
        name:
          type: string
          description: The full name of the product
          example: "Watties Baked Beans - 420g can"
        description:
          type: string
          description: A desription of the product
          example: "Baked Beans as they should be."
        manufacturer:
          type: string
          description: The manufacturer of the product
          example: "Heinz Wattie's Limited"
        recommendedRetailPrice:
          type: number
          format: double
          example: 2.20
          description: The recommended retail price of the product, given in whatever currency is in use at the business's location.
        created:
          type: string
          format: date-time
          readOnly: true
          description: The date and time when this product was created. This field is generated automatically by the server.
        images:
          type: array
          readOnly: true
          description: List of images of a product (implemented in U16).
          items:
            $ref: '#/components/schemas/Image'
...
    ListOfNotifications:
      example: [{id: 43,
                 userId: 50,
                 listingId: 30,
                 status: Expired,
                 created: '2021-04-27'
      },
        {id: 50,
         userId: 60,
         listingId: 40,
         status: Expired,
         created: '2021-05-27'
        }]
...